Remove futures-based batch verification API.
The futures-based batch verification design is still good, but it turns out
that things like the latency bounds and control of when to flush the batch
should be common across different kinds of batchable verification.  This should
be done by the forthcoming `tower-batch` middleware currently in the Zebra
repo.
