+++ src/lib.rs
@@ -6,17 +6,15 @@
 
 //! Docs require the `nightly` feature until RFC 1990 lands.
 
+mod batch;
 mod constants;
 mod error;
 mod public_key;
 mod secret_key;
 mod signature;
 
+pub use batch::BatchVerifier;
 pub use error::Error;
 pub use public_key::{PublicKey, PublicKeyBytes};
 pub use secret_key::SecretKey;
 pub use signature::Signature;
-
-#[cfg_attr(feature = "nightly", doc(cfg(feature = "batch")))]
-#[cfg(feature = "batch")]
-pub mod batch;
